% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_print_functions.R
\name{printer_funs}
\alias{printer_funs}
\alias{PrintSection}
\alias{printstamp}
\alias{printlines}
\alias{print_runtime}
\alias{buildPyramid}
\title{Print Helper}
\usage{
PrintSection(stitle = NULL, header = NULL, bullets = NULL,
  tag = "NOTE", color = "white")

printstamp(msg, sym = "#", color = c("blue", "red", "cyan", "green",
  "magenta", "yellow", "white", "silver"), face = c("bold", "italic"))

printlines(liheader = NULL, litems = NULL, likeyword = "TAG",
  lileadcol = "black", likwbgfill = "yellow", likwtxtcol = "black",
  libgfill = "blue", sym = "=", symcol = "blue", is_ordered = TRUE)

print_runtime(t, sym = "#")

buildPyramid()
}
\arguments{
\item{stitle}{Title of section for PrintSection}

\item{header}{Section header text for PrintSection}

\item{bullets}{list of notes displayed as bullets in higher level PrintSection function}

\item{tag}{arg for higher level PrintSection function. Content of tag box fill}

\item{color}{arg for higher level PrintSection function. Color of tag box fill}

\item{msg}{message message for printstamp}

\item{sym}{symbol to use to print a sectioned list}

\item{face}{functions from crayon (one of bold or italic)}

\item{liheader}{text to print as header of list}

\item{litems}{list items}

\item{likeyword}{list items keyword or tag}

\item{lileadcol}{color of the leading list item identifier}

\item{likwbgfill}{fill color of the keyword background}

\item{likwtxtcol}{text color of the keyword}

\item{libgfill}{fill color of the leading list identifiers}

\item{symcol}{color of the list partitioning}

\item{is_ordered}{TRUE (default) means the leading list identifer is numbered bullet pts}

\item{t}{time object system time for print_runtime}
}
\description{
Print Helper
}
\section{Functions}{
\itemize{
\item \code{PrintSection}: higher level function to print message on console

\item \code{printstamp}: print message on console

\item \code{printlines}: print message on console

\item \code{print_runtime}: print timing on console

\item \code{buildPyramid}: messing around with crayon
}}

