% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_print_functions.R
\name{printer_funs}
\alias{printer_funs}
\alias{printstamp}
\alias{printlines}
\alias{print_rt}
\title{Print Helper}
\usage{
printstamp(msg, sym = "#")

printlines(liheader = NULL, litems = NULL, likeyword = "TAG",
  lileadcol = "black", likwbgfill = "yellow", likwtxtcol = "black",
  libgfill = "blue", sym = "=", symcol = "blue", is_ordered = TRUE)

print_rt(t)
}
\arguments{
\item{msg}{message message for printstamp}

\item{sym}{symbol to use to print a sectioned list}

\item{liheader}{text to print as header of list}

\item{litems}{list items}

\item{likeyword}{list items keyword or tag}

\item{lileadcol}{color of the leading list item identifier}

\item{likwbgfill}{fill color of the keyword background}

\item{likwtxtcol}{text color of the keyword}

\item{libgfill}{fill color of the leading list identifiers}

\item{symcol}{color of the list partitioning}

\item{is_ordered}{TRUE (default) means the leading list identifer is numbered bullet pts}

\item{t}{time object system time for print_rt}
}
\description{
Print Helper
}
\section{Functions}{
\itemize{
\item \code{printstamp}: print message on console

\item \code{printlines}: print message on console

\item \code{print_rt}: print timing on console
}}

